+++ db4e/Widgets/NavPane.py
@@ -20,52 +20,79 @@ from db4e.Constants.Labels import (
     NEW_LABEL, P2POOL_SHORT_LABEL, XMRIG_SHORT_LABEL
 )
 from db4e.Constants.Fields import (
-    INSTANCE_FIELD, MONEROD_FIELD, MONEROD_REMOTE_FIELD, P2POOL_FIELD, XMRIG_FIELD
+    DB4E_FIELD, MONEROD_FIELD, P2POOL_FIELD, USER_WALLET_FIELD, 
+    VENDOR_DIR_FIELD
 )
 
 class NavPane(Container):
-    def __init__(self, initialized_flag: bool, config: Config, **kwargs):
-        super().__init__(**kwargs)
-        self._initialized = initialized_flag
+    def __init__(self, config: Config):
+        super().__init__()
         self.depl_mgr = DeploymentMgr(config)
+        self._initialized = False
 
         self.depls = Tree(DEPLOYMENTS_LABEL, id="tree_deployments")
         self.depls.root.add_leaf(DB4E_LABEL)
         self.depls.guide_depth = 3
         self.depls.root.expand()
 
-        self.metrics = Tree(METRICS_LABEL, id="tree_metrics")
-        self.metrics.root.expand()
-
+        #self.metrics = Tree(METRICS_LABEL, id="tree_metrics")
+        #self.metrics.root.expand()
+        #self.metrics.guide_depth = 3
         self.donations = Label(DONATIONS_LABEL, id="donations")
+        self.refresh_nav_pane()
+
+    def check_initialized(self):
+        db4e_rec = self.depl_mgr.get_deployment(DB4E_FIELD)
+        if db4e_rec and db4e_rec.get(VENDOR_DIR_FIELD) and db4e_rec.get(USER_WALLET_FIELD):
+            self.set_initialized(True)
+        else:
+            self.set_initialized(False)
 
     def compose(self) -> ComposeResult:
-        yield Vertical(self.depls, self.metrics, self.donations, id="navpane")
+        #yield Vertical(self.depls, self.metrics, self.donations, id="navpane")
+        yield Vertical(self.depls, self.donations, id="navpane")
 
-    async def on_mount(self) -> None:
-        self.metrics.guide_depth = 3
+    def is_initialized(self) -> bool:
+        return self._initialized
 
-    async def on_tree_node_selected(self, event: Tree.NodeSelected) -> None:
+    def on_tree_node_selected(self, event: Tree.NodeSelected) -> None:
         if not event.node.children:
             self.post_message(NavLeafSelected(
                 self, parent=event.node.parent.label, leaf=event.node.label))
             event.stop()
 
     def refresh_nav_pane(self) -> None:
-        if self._initialized:
-            self.depls.root.remove_children()
+        if not self.is_initialized():
+            self.check_initialized()
 
+        if self.is_initialized():
+            # Rebuild the NavPane trees
+            self.depls.root.remove_children()
             self.depls.root.add_leaf(DB4E_LABEL)
-            monero_node = self.depls.root.add(MONEROD_SHORT_LABEL)
-            p2pool_node = self.depls.root.add(P2POOL_SHORT_LABEL)
-            xmrig_node = self.depls.root.add(XMRIG_SHORT_LABEL)
             self.depls.root.expand()
 
-            m_depls = self.depl_mgr.get_deployments(MONEROD_FIELD)
-            for instance in m_depls.keys():
+            # The Monero tree
+            monero_node = self.depls.root.add(MONEROD_SHORT_LABEL)
+            instances = self.depl_mgr.get_deployment_instances(MONEROD_FIELD)
+            for instance in instances:
                 monero_node.add_leaf(instance)
             monero_node.add_leaf(NEW_LABEL)
             monero_node.expand()
+
+            p2pool_node = self.depls.root.add(P2POOL_SHORT_LABEL)
+            instances = self.depl_mgr.get_deployment_instances(P2POOL_FIELD)
+            for instance in instances:
+                p2pool_node.add_leaf(instance)
+            # Only allow P2Pool deployments if a Monero daemon deployment exists
+            if len(monero_node.children) > 1:
+                p2pool_node.add_leaf(NEW_LABEL)
+            p2pool_node.expand()
+
+            # Only display XMRig if a P2Pool deployment exists
+            xmrig_node = self.depls.root.add(XMRIG_SHORT_LABEL)
+            # Only allow XMRig deployments if a P2Pool deployment exists
+            if len(p2pool_node.children) > 1:
+                xmrig_node.add_leaf(NEW_LABEL)
 
     def set_initialized(self, value: bool) -> None:
         self._initialized = value